+++ src/components/pages/stats/LifetimeStats.vue
@@ -1,12 +1,10 @@
 <script setup lang="ts">
 import { ref, computed, onMounted } from 'vue';
 
-import { useRouter } from 'vue-router';
-const router = useRouter();
-
-import { User } from '@prisma/client';
 import { useUserStore } from 'src/stores/user.ts';
 const userStore = useUserStore();
+
+import { TALLY_MEASURE_INFO, formatCountValue, formatCountCounter } from 'src/lib/tally.ts';
 
 import { getDayCounts, DayCount } from 'src/lib/api/stats.ts';
 const dayCounts = ref<DayCount[]>([]);
@@ -17,6 +15,20 @@ async function loadDayCounts() {
 import { calculateDailyStats } from 'src/lib/stats.ts';
 const dailyStats = computed(() => {
   return calculateDailyStats(dayCounts.value);
+});
+
+const measures = computed(() => {
+  const measuresAvailable = Object.keys(dailyStats.value.days);
+  const startingBalanceMeasures = Object.keys(userStore.user.userSettings.lifetimeStartingBalance);
+  // this makes sure they're in a consistent order
+  return Object.keys(TALLY_MEASURE_INFO).filter(measure => measuresAvailable.includes(measure) || startingBalanceMeasures.includes(measure));
+});
+
+const totals = computed(() => {
+  return measures.value.reduce((obj, measure) => {
+    obj[measure] = (dailyStats.value.totals[measure] || 0) + (userStore.user.userSettings.lifetimeStartingBalance[measure] || 0);
+    return obj;
+  }, {});
 });
 
 const dayCountsByYear = computed(() => {
@@ -35,8 +47,9 @@ const dayCountsByYear = computed(() => {
 import ApplicationLayout from 'src/layouts/ApplicationLayout.vue';
 import type { MenuItem } from 'primevue/menuitem';
 import SectionTitle from 'src/components/layout/SectionTitle.vue';
+import StatTile from 'src/components/goal/StatTile.vue';
 import YearlyHeatmap from 'src/components/stats/YearlyHeatmap.vue';
-import StreakCounter from 'src/components/dashboard/StreakCounter.vue';
+// import StreakCounter from 'src/components/dashboard/StreakCounter.vue';
 
 const breadcrumbs: MenuItem[] = [
   { label: 'Stats' },
@@ -53,8 +66,15 @@ onMounted(() => {
     :breadcrumbs="breadcrumbs"
   >
     <div class="max-w-screen-lg">
-      <SectionTitle title="Lifetime Stats" />
-      <p>Many stat! Much wow.</p>
+      <SectionTitle title="Lifetime Totals" />
+      <div class="total-counts flex flex-wrap justify-evenly gap-2 mb-4">
+        <StatTile
+          v-for="measure in measures"
+          :key="measure"
+          :highlight="formatCountValue(totals[measure], measure)"
+          :suffix="formatCountCounter(totals[measure], measure)"
+        />
+      </div>
       <!-- <p><pre>{{ JSON.stringify(userStore.user.userSettings) }}</pre></p>
       <p><pre>{{ JSON.stringify(dailyStats) }}</pre></p> -->
       <div class="flex flex-col gap-4">